organize test helpers into multiple files, remove @yarnpkg/fslib dep, extract ctxTmpDir to its own context builder
