+++ package.json
@@ -122,7 +122,6 @@
     "@types/react": "^16.14.19",
     "@types/rimraf": "^3.0.0",
     "@types/semver": "^7.1.0",
-    "@yarnpkg/fslib": "^2.4.0",
     "ava": "^5.1.1",
     "axios": "^0.21.1",
     "dprint": "^0.25.0",
