+++ src/test/helpers/index.ts
@@ -1,0 +1,8 @@
+export * from './misc';
+export * from './ctx-tmp-dir';
+export * from './ctx-ts-node';
+export * from './paths';
+export * from './command-lines';
+export * from './reset-node-environment';
+export * from './version-checks';
+console.dir(Object.entries(exports).map(([k, v]) => [k, typeof v]));
